Convert the plugin to Lua
With this change we convert the plugin from Vim script to Lua. Lua seems
to be the new "native" language for writing plugins and, while still
questionable, seems less cryptic than Vim script. The conversion is done
in a fairly mechanical manner -- more work is necessary to switch over
to more typed APIs and avoid usage of vim.api.nvim_input() where
possible.