Work with a proper scratch buffer
So far we have used a rather ad hoc approach to creating the buffer in
which we present the list of available buffers to the user, relying on a
user command etc. The overall approach was not great, because it could
easily confuse the users when they wanted to just close the editor
instance, but this unsaved hidden buffer would prevent that from
happening.
With this change we rework the approach we take, creating and working
with a proper scratch buffer that gets destroyed when it is being hidden
instead, which mitigates this problem to a large degree.