Set alternate buffer to previously active one
When the user selects a buffer, our scratch buffer effectively acts as
the alternative. Because that buffer vanishes on hiding, no alternate
ends up being available in the end.
From a user's point of view, the scratch buffer should be entirely
transparent and irrelevant and what matters to them is the previously
active buffer. Luckily we still have that one as the "old" alternate and
so with this change we make sure to switch to it before honoring the
user's new selection. This way, said buffer is preserved as the
alternate moving forward as well.